requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pillow==10.0.1
pdf2image==1.16.3
pandas==2.1.4
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0